requirements.txt
Django==4.2.10
djangorestframework==3.14.0
python-dotenv==1.0.0
celery==5.3.6
django-celery-results==2.5.1
django-celery-beat==2.5.0
spacy==3.7.2
sentence-transformers==2.2.2
redis==5.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.5
markdown==3.5
pandas==2.1.3
pymupdf==1.23.6
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0 